move editor to MonacoEditor
